Move helper to helpers folder
